Fixed flickering hidden buttons (#16)
* Fixed updateVisibility flicker on hide

When control button hid key buttons, the opacity would hit zero but this._showing was never set false. That would ause the esle if branch on the next run of updateVisibility to set opacity to a positive amount and the keybutton would begin flickering since this._hiding is was still === true. I've corrected this by setting this._showing = false; when opacity <= 0 in updateVisibility;

* Fixed updateVisibility flicker on hide

When control button hid key buttons, the opacity would hit zero but this._showing was never set false. That would ause the esle if branch on the next run of updateVisibility to set opacity to a positive amount and the keybutton would begin flickering since this._hiding is was still === true. I've corrected this by setting this._showing = false; when opacity <= 0 in update visibility;